add tests for numpy view usage

+++ pymc3/variational/opvi.py
@@ -741,7 +741,7 @@ class Approximation(object):
                                             tt.as_tensor(_shape)])
                 else:
                     shape = np.concatenate([space.shape[:-1],
-                                            _shape])
+                                            _shape]).astype(int)
 
             else:
                 shape = space.shape[:-1]
